deepfake-patch-audit/deployment/requirements.txt
# Raspberry Pi Flask Server Requirements

# Web Framework
Flask==2.3.2
Werkzeug==2.3.6

# ONNX Runtime (CPU inference on Raspberry Pi)
onnxruntime==1.15.0

# Image processing
Pillow==10.0.0

# HTTP requests (for webhook alerts)
requests==2.31.0

# JSON processing (optional, for better JSON handling)
# Already built-in to Python, but listed for reference

# Email support (for email alerts)
# Already built-in to Python (smtplib)

# Optional: For development and testing
# numpy is automatically installed with onnxruntime

# Note:
# - onnxruntime-rpi should be used if available for ARM optimization
# - Some versions may need adjustment based on Raspberry Pi OS version
# - Consider using onnxruntime-silicon if running on newer ARM-based systems
